pyproject.toml
[project]
name = "BBDecoder"
version = "1.1.12"
authors = [
    {name = "Sharjeel Masood", email = "sharjeelmasood.sm@gmail.com"},
    {name = "Saeed Ahmed", email = "saeedahmad.icp@gmail.com"}
]
maintainers = [
    {name = "Sharjeel Masood", email = "sharjeelmasood.sm@gmail.com"},
    {name = "Saeed Ahmed", email = "saeedahmad.icp@gmail.com"}
]
description = "Analysis of Pytorch Models"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

keywords = [
    "pytorch", "visualization", "analysis",
    "torch", "deep learning", "machine learning",
    "ml", "neural network"
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "torchview>=0.2.7",
    "graphviz"
]

[project.urls]
homepage = "https://github.com/sharjeel1999/Black-Box-Decoder"
repository = "https://github.com/sharjeel1999/Black-Box-Decoder"


[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
packages = [
    "BBDecoder",
    "BBDecoder.analysis",
    "BBDecoder.utilities",
    "BBDecoder.visualizer",
    "BBDecoder.wrappers",
    ]

license-files = []
